# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for authenticated Linux users to manage Containerlab deployments, topology files, node access, captures, and host networking tools.\n\nMost endpoints under /api/v1 require a JWT from POST /login. Send it as: Authorization: Bearer \u003ctoken\u003e.",
        "title": "Containerlab API Server",
        "contact": {
            "name": "Containerlab API Server maintainers",
            "url": "https://github.com/srl-labs/clab-api-server/issues"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/generate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates a containerlab topology from CLOS definitions and can optionally deploy it for the authenticated user.\n\n**Notes**\n- Deployment is denied if a lab with the target name already exists.\n- The `images` and `licenses` fields map node kind to image or license path (e.g., {\"nokia_srlinux\":\"ghcr.io/...\"}).\n- When `deploy=true`, the topology is saved to the user's managed lab directory and `outputFile` is ignored.\n- When `deploy=false` and `outputFile` is empty, YAML is returned in the response.\n- When `deploy=false` and `outputFile` is set, the relative file path is saved inside the authenticated user's managed workspace.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Topology Generation"
                ],
                "summary": "Generate containerlab topology",
                "parameters": [
                    {
                        "description": "Topology generation parameters. The 'images' field maps kind to image path.",
                        "name": "generate_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GenerateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Generation successful (YAML or deploy output)",
                        "schema": {
                            "$ref": "#/definitions/models.GenerateResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input parameters",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict (Lab already exists and Deploy=true)",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error or clab execution failed",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/labs/{labName}/nodes/{nodeName}/logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns logs for a lab node.\n\n**Notes**\n- When `follow=true`, the response streams as NDJSON (one JSON object per line) until the client disconnects or the 30-minute timeout.",
                "produces": [
                    "application/json",
                    "application/x-ndjson"
                ],
                "tags": [
                    "Logs"
                ],
                "summary": "Get node logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the lab",
                        "name": "labName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Full name of the container (node)",
                        "name": "nodeName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "all",
                        "description": "Number of lines to show from the end of logs (default all). Use an integer or 'all'.",
                        "name": "tail",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Follow log output (stream logs as NDJSON). Note: In Swagger UI, streaming may not display correctly.",
                        "name": "follow",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Container logs (follow=false). When follow=true, response is NDJSON stream of LogLine objects.",
                        "schema": {
                            "$ref": "#/definitions/models.LogsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input (lab name, node filter, etc.)",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (not owner of the lab)",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Lab or node not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid credentials or user not in allowed group"
                }
            }
        },
...
        "models.GenerateNodeTier": {
            "type": "object",
            "required": [
                "count"
            ],
            "properties": {
                "count": {
                    "description": "Number of nodes in this tier",
                    "type": "integer",
                    "minimum": 1,
                    "example": 4
                },
                "kind": {
                    "description": "Node kind (defaults to 'srl'/'nokia_srlinux' if omitted)",
                    "type": "string",
                    "example": "nokia_srlinux"
                },
                "type": {
                    "description": "Node type within the kind",
                    "type": "string",
                    "example": "ixrd3"
                }
            }
        },
        "models.GenerateRequest": {
            "type": "object",
            "required": [
                "name",
                "tiers"
            ],
            "properties": {
                "defaultKind": {
                    "type": "string",
                    "example": "nokia_srlinux"
                },
                "deploy": {
                    "type": "boolean",
                    "example": true
                },
                "groupPrefix": {
                    "type": "string",
                    "example": "clos-tier"
                },
                "images": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "nokia_srlinux": "ghcr.io/nokia/srlinux:latest"
                    }
                },
                "ipv4Subnet": {
                    "type": "string",
                    "example": "172.20.20.0/24"
                },
                "ipv6Subnet": {
                    "type": "string",
                    "example": "2001:172:20:20::/64"
                },
                "licenses": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "nokia_srlinux": "/opt/licenses/srl.lic"
                    }
                },
                "managementNetwork": {
                    "type": "string",
                    "example": "clos-mgmt"
                },
                "maxWorkers": {
                    "type": "integer",
                    "example": 4
                },
                "name": {
                    "type": "string",
                    "example": "3-tier-clos"
                },
                "nodePrefix": {
                    "type": "string",
                    "example": "clos-node"
                },
                "outputFile": {
                    "type": "string",
                    "example": "clos.yml"
                },
                "tiers": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/models.GenerateNodeTier"
                    }
                }
            }
        },
        "models.GenerateResponse": {
            "type": "object",
            "properties": {
                "deployOutput": {
                    "description": "The output from the deploy command (only if Deploy=true). JSON object keyed by lab name.\nUse swaggertype:\"object\" to represent json.RawMessage in Swagger.",
                    "type": "object"
                },
                "message": {
                    "description": "Message indicating success or failure.",
                    "type": "string"
                },
                "savedFilePath": {
                    "description": "Path where the file was saved (if Deploy=true, it's the path in the user's managed lab directory; if Deploy=false, it's the OutputFile path if provided).",
                    "type": "string"
                },
                "topologyYaml": {
                    "description": "The generated topology YAML (only if Deploy=false and OutputFile is empty).",
                    "type": "string"
                }
            }
        },
...
        "models.LogsResponse": {
            "type": "object",
            "properties": {
                "containerName": {
                    "type": "string"
                },
                "logs": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token. Example: \"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}